pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trellis-datamodel"
version = "0.6.0"
description = "Visual data model editor for dbt projects"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Tim Hiebenthal"},
]
keywords = ["dbt", "data-modeling", "erd", "data-engineering", "analytics-engineering", "visualization", "schema"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "dbt-core>=1.10.5,<2.0",
    "dbt-colibri>=0.1.0",
    "dbt-duckdb>=1.10.0",
    "fastapi>=0.121.3",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "ruamel.yaml>=0.18.0",
    "typer>=0.9.0",
    "uvicorn>=0.38.0",
]

[project.urls]
Homepage = "https://app.capacities.io/home/8b7546f6-9028-4209-a383-c4a9ba9be42a"
Repository = "https://github.com/timhiebenthal/trellis-datamodel"
Issues = "https://github.com/timhiebenthal/trellis-datamodel/issues"

[project.scripts]
trellis = "trellis_datamodel.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
dbt-example = [
    "dbt-duckdb==1.10",
    "duckdb>=1.4.2",
    "faker>=24.0.0",
    "marimo>=0.18.0",
    "nba-api>=1.11.3",
    "pandas>=2.3.3",
    "tqdm>=4.67.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["trellis_datamodel*"]

[tool.setuptools.package-data]
trellis_datamodel = ["static/**/*"]

[tool.pytest.ini_options]
testpaths = ["trellis_datamodel/tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.1",
]
